requirements_dashboard.txt
# Dashboard-specific dependencies
plotly==5.17.0
dash==2.14.2
dash-leaflet==0.1.23
dash-bootstrap-components==1.5.0
folium>=0.14.0

# Core scientific computing (already in main requirements)
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Geospatial libraries (already in main requirements)
geopandas>=0.10.0
rasterio>=1.2.0
pyproj>=3.2.0
shapely>=1.8.0

# Visualization (already in main requirements)
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities (already in main requirements)
pyyaml>=6.0

# Statistical analysis (already in main requirements)
scikit-learn>=1.0.0
statsmodels>=0.13.0